refactor(oc-docs): align Topbar to per-component folder convention (BRU-3572)
Structure + naming only — no behavior, props, or visual change. Adopts the
house convention from PR #45 (each component a folder with <Name>.tsx +
<Name>.spec.tsx + StyledWrapper.ts + index.ts; pure utils in src/utils;
hooks in src/hooks; shared glyphs in src/assets/icons.tsx).

Moves:
- Promote shared primitives to top-level src/components/: IconButton/,
  InitialsAvatar/, OpenInBrunoButton/ (reusable across lanes).
- Keep Topbar/, Topbar/Brand/, Topbar/MobileOverflow/ as their own folders.
- Split the single Topbar/StyledWrapper.ts into per-component StyledWrapper.ts.
- Pure utils -> src/utils: buildBrunoDeepLink, getInitials (+ .spec.ts).
- Hooks -> src/hooks: useTopbarLayout, useCanRunBrunoApp (+ .spec.ts).
- Raw glyphs -> src/assets/icons.tsx (SearchIcon, HamburgerIcon, OverflowIcon,
  BrunoGlyph). Matches PR #45's icons.tsx style/baseIconProps; see the file's
  reconciliation note for the (mechanical) merge with #45's glyph set.
- *.test.ts -> *.spec.ts(x); add a colocated render spec per component.

IconButton consolidation:
- Single canonical src/components/IconButton/ (icon-only button primitive,
  aria-label required). Removed the duplicate, unused IconButton from
  ui/MinimalComponents.tsx (the other exports there are untouched) and the
  local copy in Topbar/icons.tsx. All callers import the shared path.

Dev scaffolding removed from shipped src/ + package root: deleted
src/topbarHarness.tsx, src/deviceCheck.tsx, topbar-harness.html,
topbar-device-check.html. The e2e harness is now gated behind dev.tsx
(?view=topbar-harness); e2e navigates there instead of the html.

Public API unchanged: Topbar still exports from src/components/Topbar
(index.ts re-export) for BRU-3188. OpenCollection updated to import
buildBrunoDeepLink from src/utils.
