+++ packages/oc-docs/src/components/InitialsAvatar/index.ts
@@ -1,0 +1,2 @@
+export { default as InitialsAvatar } from './InitialsAvatar';
+export type { InitialsAvatarProps } from './InitialsAvatar';
