+++ packages/oc-docs/src/components/OpenCollection/OpenCollection.tsx
@@ -7,7 +7,7 @@ import type { OpenCollection as IOpenCollection } from '@opencollection/types';
 import PlaygroundDrawer from '../PlaygroundDrawer/PlaygroundDrawer';
 import Docs from '../Docs/Docs';
 import { Topbar } from '../Topbar';
-import { buildBrunoDeepLink } from '../Topbar/buildBrunoDeepLink';
+import { buildBrunoDeepLink } from '../../utils/buildBrunoDeepLink';
 import { parseYaml } from '../../utils/yamlUtils';
 import { hydrateWithUUIDs } from '../../utils/items';
 import { getItemType, isFolder } from '../../utils/schemaHelpers';
