+++ packages/oc-docs/src/components/OpenInBrunoButton/index.ts
@@ -1,0 +1,2 @@
+export { default as OpenInBrunoButton } from './OpenInBrunoButton';
+export type { OpenInBrunoButtonProps } from './OpenInBrunoButton';
