+++ packages/oc-docs/src/components/Topbar/Brand/index.ts
@@ -1,0 +1,3 @@
+export { default } from './Brand';
+export { default as Brand } from './Brand';
+export type { BrandProps } from './Brand';
