+++ packages/oc-docs/src/components/Topbar/MobileOverflow/index.ts
@@ -1,0 +1,3 @@
+export { default } from './MobileOverflow';
+export { default as MobileOverflow } from './MobileOverflow';
+export type { MobileOverflowProps } from './MobileOverflow';
