+++ packages/oc-docs/src/components/Topbar/index.ts
@@ -1,23 +1,2 @@
 export { default as Topbar } from './Topbar';
 export type { TopbarProps } from './Topbar';
-export { default as OpenInBrunoButton } from './OpenInBrunoButton';
-export type { OpenInBrunoButtonProps } from './OpenInBrunoButton';
-export { default as Brand } from './Brand';
-export type { BrandProps } from './Brand';
-export { default as InitialsAvatar } from './InitialsAvatar';
-export type { InitialsAvatarProps } from './InitialsAvatar';
-export { IconButton } from './icons';
-export type { IconButtonProps } from './icons';
-export { getInitials } from './getInitials';
-export { buildBrunoDeepLink } from './buildBrunoDeepLink';
-export {
-  useTopbarLayout,
-  layoutModeForWidth,
-  showsHamburger,
-  type TopbarLayoutMode,
-} from './useTopbarLayout';
-export {
-  useCanRunBrunoApp,
-  computeCanRunBrunoApp,
-  type DeviceEnv,
-} from './useCanRunBrunoApp';
