+++ packages/oc-docs/src/dev.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-refresh/only-export-components -- dev-only entry, not a component module */
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
@@ -12,6 +13,7 @@ import 'prismjs/components/prism-xml-doc';
 import 'prismjs/components/prism-http';
 import 'prismjs/components/prism-graphql';
 import OpenCollection from './components/OpenCollection/OpenCollection';
+import { Topbar } from './components/Topbar';
 import { createOpenCollectionStore } from './store/store';
 import { sampleCollectionYaml } from './sampleCollection';
 
@@ -19,7 +21,6 @@ import { sampleCollectionYaml } from './sampleCollection';
 if (typeof window !== 'undefined') {
   (window as any).Prism = Prism;
 }
-
 
 // Development App component
 const DevApp: React.FC = () => {
@@ -37,11 +38,62 @@ const DevApp: React.FC = () => {
   );
 };
 
+/**
+ * Dev-only Topbar harness (gated behind `?view=topbar-harness`). Mounts the
+ * Topbar with both slots filled so the e2e suite can exercise the search-expand
+ * and overflow popover, which need the search / env-switcher slot nodes that
+ * downstream tickets supply. Not part of the shipped standalone build.
+ */
+const SearchSlot: React.FC = () => (
+  <input
+    data-testid="search-slot-input"
+    placeholder="Search endpoints…"
+    style={{
+      width: '100%',
+      height: '28px',
+      boxSizing: 'border-box',
+      padding: '5px 8px',
+      border: '1px solid var(--oc-border-border1)',
+      borderRadius: 'var(--oc-border-radius-base)',
+      background: 'var(--oc-background-base)',
+      color: 'var(--oc-text)',
+      fontFamily: 'var(--font-sans)',
+    }}
+  />
+);
+
+const EnvSwitcherSlot: React.FC = () => (
+  <div data-testid="env-switcher-slot" style={{ display: 'flex', gap: '8px' }}>
+    <span>Show vars</span>
+    <span>Development</span>
+  </div>
+);
+
+const TopbarHarness: React.FC = () => (
+  <div style={{ minHeight: '200vh' }}>
+    <Topbar
+      collectionName="Hotel Booking API"
+      version="v1.0.0"
+      searchSlot={<SearchSlot />}
+      envSwitcherSlot={<EnvSwitcherSlot />}
+      openInBrunoHref="bruno://app/collection/import/git?url=https%3A%2F%2Fexample.com%2Frepo.git"
+      onToggleSidebar={() => {
+        (window as unknown as { __toggleSidebarCalls?: number }).__toggleSidebarCalls =
+          ((window as unknown as { __toggleSidebarCalls?: number }).__toggleSidebarCalls ?? 0) + 1;
+      }}
+    />
+    <div style={{ padding: '24px' }}>Scroll content to verify the bar stays pinned.</div>
+  </div>
+);
+
+const view =
+  typeof window !== 'undefined' ? new URLSearchParams(window.location.search).get('view') : null;
+
 // Render the app
 const container = document.getElementById('root');
 if (container) {
   const root = createRoot(container);
-  root.render(<DevApp />);
+  root.render(view === 'topbar-harness' ? <TopbarHarness /> : <DevApp />);
 } else {
   console.error('Root container not found');
 }
