+++ packages/oc-docs/src/hooks/index.ts
@@ -1,1 +1,8 @@
 export { useMarkdownRenderer } from './useMarkdownRenderer';
+export {
+  useTopbarLayout,
+  layoutModeForWidth,
+  showsHamburger,
+  type TopbarLayoutMode,
+} from './useTopbarLayout';
+export { useCanRunBrunoApp, computeCanRunBrunoApp, type DeviceEnv } from './useCanRunBrunoApp';
