+++ packages/oc-docs/src/utils/getInitials.spec.ts
