+++ packages/oc-docs/src/utils/getInitials.ts
