--- packages/oc-docs/topbar-harness.html
@@ -1,12 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Topbar Harness</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/topbarHarness.tsx"></script>
-  </body>
-</html>
